LPS-158624 Move XMLUtil.stripInvalidChars into portal-kernel since it does not depend on 3rd party library
